Struct plonk_jubjub::elgamal::ElgamalCipher

pub struct ElgamalCipher { /* fields omitted */ }

Tuple for assymetric encryption using ElGamal algorithm.

Example

use dusk_jubjub::elgamal::ElgamalCipher;
use dusk_jubjub::{JubJubScalar, GENERATOR_EXTENDED};

fn main() {
    // Bob's (sender) secret and message
    let bob_secret = JubJubScalar::random(&mut rand::thread_rng());
    let message = JubJubScalar::random(&mut rand::thread_rng());
    let message = GENERATOR_EXTENDED * message;

    // Alice's (receiver) secret and public
    let alice_secret = JubJubScalar::random(&mut rand::thread_rng());
    let alice_public = GENERATOR_EXTENDED * alice_secret;

    let cipher = ElgamalCipher::encrypt(
        &bob_secret,
        &alice_public,
        &GENERATOR_EXTENDED,
        &message,
    );
    let decrypt = cipher.decrypt(&alice_secret);

    assert_eq!(message, decrypt);
}

Let p and G = α be defined by the parameters of JubJub.

Let a be Alice’s secret, and A = G · a

Let b be Bob’s secret, and B = G · b

Encryption

Bob should do the following:

Obtain Alice’s authentic public key A.

Represent the message M as a point of JubJub defined such as M = G ·m where m is a scalar in JubJubScalar.

Compute γ = G · b and δ = M + (A ·b).

Send the ciphertext c = (γ; δ) to Alice.

Decryption

To recover plaintext M from c, Alice should do the following:

Recover M by computing δ - γ · a.

Homomorphism

A function f is homomorphic when f(a · b) = f(a) · f(b).

This implementation extends the homomorphic property of ElGamal to addition, subtraction and multiplication.

The addition and subtraction are homomorphic with other ElgamalCipher structures.

The multiplication is homomorphic with JubJubScalar scalars.

Being E the encrypt and D the decrypt functions, here follows an example: D{E[x * (a + b)]} == D{x * [E(a) + E(b)]}

Implementations

impl ElgamalCipher

pub fn new(gamma: JubJubExtended, delta: JubJubExtended) -> Self

ElgamalCipher constructor

pub fn gamma(&self) -> &JubJubExtended

Getter for the gamma public key

pub fn delta(&self) -> &JubJubExtended

Getter for the delta ciphertext

pub fn encrypt(
secret: &JubJubScalar,
public: &JubJubExtended,
generator: &JubJubExtended,
message: &JubJubExtended
) -> Self

Uses assymetric encryption to return a cipher construction.

The decryption will expect the secret of public.

pub fn decrypt(&self, secret: &JubJubScalar) -> JubJubExtended

Perform the decryption with the provided secret.